Grep for more evil in test as a signature can contain evil
The test was failing for me as the PGP signature contained "GeviljGB"
which is clearly not the evil we are looking for. Sadly, reproducing
this is hard, but random failures of the test suite are annoying, so
lets grep for a bit longer strings that can't happen in the signature
block to avoid this particular problem.