Fix regression that prevented sending intermediate client certificates
The change from the GNUTLS backend to the OpenSSL backend only loads
the first certificate from a client certificate file. As a result it
is no longer possible to send intermediate client certificates to the
server to assist in certificate verification.  Instead call
ssl_ctx_use_certificate_chain_file so intermediate certificates can be
sent.

Signed-off-by: Sam Hartman <hartmans@debian.org>