solver3: Support comparing packages and versions
When comparing packages, we compare the best version for the
package. To determine the best version, the rules for comparisons
are applied normally.

This allows us to mix packages and versions in a clause as the
solutions for a dependency, which means we'll be able to defer
the selection of a particular version of a package to a later
time.