solver3: Use a package clause for optional roots
Instead of iterating over the version here and picking it, just
enqueue the package as well, which should allow us to select the
version at a later time.

This also causes a funny inverse problem now, though, as was evidenced
in one of the test cases: To summarize, if our optional roots are all
single items, they will be considered soft-unit, causing them to be
processed in order.

However it can be that an optional root has a specific version
selected because another version was rejected. Consider

    X Conflicts A (= 1)
    A, B have 2 versions: '2' available, '1' installed
    B (= n) Depends A (= n)

Run `apt install X`. The expected result is for A and B to be
upgraded to version 2. With only a package root, if B appears
in the cache before A however, we will get:

    Install X
        Reject A (= 1)
    Install B
        Install B (= 1)     # keep it installed
            Reject A (= 2)
    => A is being removed as both versions are rejected

Hence we do also need to re-introduce the additional version
clause, now we get:

    Install X
        Reject A (= 1)
    Install A (= 2)         # it got "promoted" to a 'stronger' soft-unit
    Install B
        Fail B (= 1)        # keep it installed
        Install B (= 2)

Introduce a root state to hold all the clauses that don't have
another owner.

moo