test: Fix skipping test cases
The `--skip` option was introduced in commit
1fbdc04be7547b8c4bb84f400329415ea0bc60b5, and it was broken straight
from the start for the case where `TESTTORUN` is not set (see last hunk
in this commit, the mistake is obvious).

It went unnoticed as the GitLab CI runs tests with the `-j` option, so
`TESTTORUN` is set. While ci.debian.net runs without -j, but doesn't
skip any test, so it doesn't trigger the bug either.