solver3: Fix error stack handling
Pushing the stack in push and popping it in pop did not
really work correctly and is more complex than needed.

Instead, push the error stack at the start of the Solve()
method and revert at the end, such that we leave exactly
at the same error stack level we entered.

To handle error clearing on backtracking, just discard any
pending errors.