apt-ftparchive: Do not output locale dependent dsc file size
Removal of std::to_string avoids the temporary std::string,
but it leads to locale-dependent output that parsers (including ours)
are not prepared to encounter especially as number formatting
can be highly ambiguous due to different rules surrounding '.' and ','.

In the distant future (c++20) this could be replaced with a variant of
std::format, which incidentally is also what std::to_string becomes
an alias for in an even more distant future (c++26).

Regression-of: 729a5c8f4fb8c841497d26170e1a47c41617ecd9
Reported-By: Stefan Lippers-Hollmann and Kel Modderman on IRC