test-apt-ftparchive-cachedb: Avoid race condition in test case
apt-ftparchive skips regenerating Contents if they are newer than
the Packages files. Most of the time, the test ended up generating
Contents and Packages files with the same-second timestamp; so we
ended up regenerating them and passing the test.

However, sometimes, the timestamp of the Contents file was 1s
higher and we ended up not regenerating the Contents file and
hence failed the test case.

Ensure the timestamps actually are the same to avoid the race
condition and make this stable.