solver3: Do not install new packages for Suggests
As a followup to the previous commit, restore the exact behavior
of the old solver to break the Suggests on upgrade here by erasing
any choices not previously installed.

This ensures that the Suggests never installs any new package, but
it will still retain previously installed packages when solver3 is
doing autoremove.