solver3: Consider more upgrades more costly when not upgrading
We inadvertently considered more upgrades as a better solution in
all cases, but if we are *installing* rather than upgrading, we
should count upgrades as a negative.

Change the code to accomodate that. This requires a few more
tricks:

- We remove the keep count check. This does not appear to work
  correctly and would have prevented the fix. We are already
  counting all possible changes, so why bother.

- We need to subtract the upgrade count from the install count
  (upgrades are a class of installs) since we now allow comparing
  solutions with different upgrade counts for install counts.

- After comparing the "new install count", we now compare upgrade
  count too, in the opposite direction of upgrading.

To test, we abuse that solver 3.0 does not upgrade all packages
in a source package when using the install command.

LP: #2105395