debian/NEWS: Document new requirements on trusted.gpg.d
All files in apt directories must follow the run-parts requirements,
the only allowed characters are:

    a-z, A-Z, 0-9, _, -, :

and '.' if an extension is required, due to the use of the general
GetFilesInDirectory() function.

Other characters were previously supported as the directory was not
enumerated by apt, but by apt-key, an 850ish line shell script using

  find "$TRUSTEDPARTS" -mindepth 1 -maxdepth 1 \( -name '*.gpg' -o -name '*.asc' \)

which obviously does not share the same requirements.

Document the difference in debian/NEWS, as it seems ill-advised to
introduce a flag-carrying version of GetFilesInDirectory() for the
sole purpose of a deprecated signing mechanism.

Closes: #1104320