solver3: Handle failed assumption in Pop()
If an Assume() failed, we will have pushed the choices stack,
but not actually pushed anything to the solved stack.

This would cause a segmentation fault trying to read the referenced
entry that doesn't actually exists, namely solved[solved.size()].

Handle this by exiting early as there's nothing to do except
pop the entry back out of the array.