fix compiling with musl
In case apt is compiled against musl-libc, then compilation fails with the
following error:

cmdline/apt-internal-solver.cc:122:20: error: 'basename' was not declared in this scope; did you mean 'rename'?
|   122 |         if (strcmp(basename(argv[0]), "solver3") == 0)
|       |                    ^~~~~~~~
|       |                    rename

To fix it, include libgen header where musl can find this function also.

Once this is fixed, compilation fails once again with musl-libc, because
the basename function takes *char argument instead of const *char:
this is because the musl implementation cuts off the trailing slashes
of the path, in case the argument is a folder path.

To account for this, instead of passing argv[0] directly, create a non-const
string copy of it, and pass that to base.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>