ExecFork: Use close_range() if available
Use close_range() if we have it in preference to iterating
over /proc/self/fd and falling back to closing all possible
fds.

This builds sets of ranges to close that take into account
the APT::Keep-Fds vector, and a test case is provided to
ensure correctness of the splitting logic.

Preference is given to close_range() over /proc as in the
optimal case, this results in a single system call!