Revert "depcache: Allow changing multi-arch versions for removed siblings"
This reverts commit 74636b64a7ee56637a3ce4ddc1d060ca9f8b77e1.

Adjust the test case accordingly, as it is no longer possible to
install libgcc-s1:s390x. We need to sort this out in the ToDepCache
translation layer at a later time.

Sadly the depcache rejects valid solutions like this; solving this
isn't trivial. Technically we need to split up the FromUser and
auto-bit handling to solve this, but even then, MarkDelete() with
FromUser=0 moves the auto-bit on metapackages; this doesn't happen
when doing FromUser=1, so we can't always do FromUser=1.

Fixes: a3fca052ca21ad222ac7f2fdd7f3fe84b44beb60